feat: redesign web ui session workspace (#185)
